This is the second version of the firefox plugin.  It is not complete at this stage.  The goal of this rewrite is to do as much logic as possible in javascript.  The notary code should interact with the rest of the extension through the queryNotaries function located in notaries.js (currently this is just a dummy implementation).  The current implementation pretends to contact the notaries and displace a green check if the pretend notaries respondend favorably and a red check otherwise.  The pretend notaries always return a 0 quorum for websites containing the string "pnc".
Things to do: Need to be able to set user preferences (how large is an acceptable quorum and how long a duration is necessary etc)

Need to actually contact the notaries.

Need to overide firefox3's draconian ssl policy (e.g. the big scary page it shows on self signed websites)

Known bugs: 
before the user has opened his first tab the code fails to receive location update events from the browser's progress listener.  I think this is a bug with firefox 3 RC2 and I will see if the problem still exists in RC3.  If it doesn't exist in 
RC3 i will figure out if its a firefox bug or a bug in the extension.





git-svn-id: https://moo.cmcl.cs.cmu.edu/svn/multissh/notarycode@228 4bf60a73-bb1f-0410-bc46-cb087b644473