Added a new test script which fetches each entry in the database once and sanity checks it (checks if it returns <server>.*</server> thats all).
Moved both test scripts to a test_scripts directory

Made notary_http not print debugging information by default and Changes a
constant to a #define:

Known Bug:
If the entry in the database is too long (e.g. citibank) the server hangs
up without responding.  I need to migrate to string_buffer to fix this.  


git-svn-id: https://moo.cmcl.cs.cmu.edu/svn/multissh/notarycode@414 4bf60a73-bb1f-0410-bc46-cb087b644473