Working on the "Report Attack" functionality.
Currently if you go to a website which perspectives is not able to validate and
click on the "Report Attack" button the plugin will pop up a window with the
data we plan to push to the networknotary servers.  I am committing it now so
that Dan can see the data I intend to push and the UI.

To Test the code, empty the whitelist.json file and go to one of the whitelisted
websites.  Perspectives will fail to validate it and you should see the UI.

Still Todo:
Minor UI tweaks: It's mostly there but needs some small stuff

Post The data: The extension code is 90% there, need to finish that up and set
up a sql table on the server to collect the data.

git-svn-id: https://moo.cmcl.cs.cmu.edu/svn/multissh/notarycode@453 4bf60a73-bb1f-0410-bc46-cb087b644473